# perf: keep a running power for scale